Refactor the path
